+++ cmd/otel-allocator/allocation/allocatortest.go
@@ -39,7 +39,7 @@ func MakeNNewTargets(n int, numCollectors int, startingIndex int) map[string]*ta
 			"i":         model.LabelValue(strconv.Itoa(i)),
 			"total":     model.LabelValue(strconv.Itoa(n + startingIndex)),
 		}
-		newTarget := target.NewItem(fmt.Sprintf("test-job-%d", i), "test-url", label, collector)
+		newTarget := target.NewItem(fmt.Sprintf("test-job-%d", i), fmt.Sprintf("test-url-%d", i), label, collector)
 		toReturn[newTarget.Hash()] = newTarget
 	}
 	return toReturn
@@ -53,6 +53,19 @@ func MakeNCollectors(n int, startingIndex int) map[string]*Collector {
 			Name:       collector,
 			NumTargets: 0,
 		}
+	}
+	return toReturn
+}
+
+func MakeNNewTargetsWithEmptyCollectors(n int, startingIndex int) map[string]*target.Item {
+	toReturn := map[string]*target.Item{}
+	for i := startingIndex; i < n+startingIndex; i++ {
+		label := model.LabelSet{
+			"i":     model.LabelValue(strconv.Itoa(i)),
+			"total": model.LabelValue(strconv.Itoa(n + startingIndex)),
+		}
+		newTarget := target.NewItem(fmt.Sprintf("test-job-%d", i), fmt.Sprintf("test-url-%d", i), label, "")
+		toReturn[newTarget.Hash()] = newTarget
 	}
 	return toReturn
 }