+++ controllers/reconcile_test.go
@@ -17,6 +17,7 @@ package controllers_test
 import (
 	"context"
 	"testing"
+	"time"
 
 	routev1 "github.com/openshift/api/route/v1"
 	"github.com/stretchr/testify/assert"
@@ -27,14 +28,20 @@ import (
 	v1 "k8s.io/api/core/v1"
 	networkingv1 "k8s.io/api/networking/v1"
 	policyV1 "k8s.io/api/policy/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/intstr"
+	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/tools/record"
 	controllerruntime "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	k8sconfig "sigs.k8s.io/controller-runtime/pkg/client/config"
+	"sigs.k8s.io/controller-runtime/pkg/manager"
 	k8sreconcile "sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	"github.com/open-telemetry/opentelemetry-operator/apis/v1alpha1"
 	"github.com/open-telemetry/opentelemetry-operator/controllers"
+	"github.com/open-telemetry/opentelemetry-operator/internal/autodetect/openshift"
 	"github.com/open-telemetry/opentelemetry-operator/internal/config"
 	"github.com/open-telemetry/opentelemetry-operator/internal/manifests"
 	ta "github.com/open-telemetry/opentelemetry-operator/internal/manifests/targetallocator/adapters"
@@ -94,6 +101,9 @@ func TestOpenTelemetryCollectorReconciler_Reconcile(t *testing.T) {
 	updatedRouteParams.OtelCol.Spec.Ingress.Type = v1alpha1.IngressTypeRoute
 	updatedRouteParams.OtelCol.Spec.Ingress.Route.Termination = v1alpha1.TLSRouteTerminationTypeInsecure
 	updatedRouteParams.OtelCol.Spec.Ingress.Hostname = expectHostname
+	deletedParams := paramsWithMode(v1alpha1.ModeDeployment)
+	now := metav1.NewTime(time.Now())
+	deletedParams.OtelCol.DeletionTimestamp = &now
 
 	type args struct {
 		params manifests.Params
@@ -488,8 +498,31 @@ func TestOpenTelemetryCollectorReconciler_Reconcile(t *testing.T) {
 				},
 			},
 		},
+		{
+			name: "collector is being deleted",
+			args: args{
+				params:  deletedParams,
+				updates: []manifests.Params{},
+			},
+			want: []want{
+				{
+					result: controllerruntime.Result{},
+					checks: []check{
+						func(t *testing.T, params manifests.Params) {
+							o := v1alpha1.OpenTelemetryCollector{}
+							exists, err := populateObjectIfExists(t, &o, namespacedObjectName(naming.Collector(params.OtelCol.Name), params.OtelCol.Namespace))
+							assert.NoError(t, err)
+							assert.False(t, exists) // There should be no collector anymore
+						},
+					},
+					wantErr:     assert.NoError,
+					validateErr: assert.NoError,
+				},
+			},
+		},
 	}
 	for _, tt := range tests {
+		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			testContext := context.Background()
 			nsn := types.NamespacedName{Name: tt.args.params.OtelCol.Name, Namespace: tt.args.params.OtelCol.Namespace}
@@ -501,13 +534,21 @@ func TestOpenTelemetryCollectorReconciler_Reconcile(t *testing.T) {
 				Config: config.New(
 					config.WithCollectorImage("default-collector"),
 					config.WithTargetAllocatorImage("default-ta-allocator"),
+					config.WithOpenShiftRoutesAvailability(openshift.RoutesAvailable),
 				),
 			})
+
 			assert.True(t, len(tt.want) > 0, "must have at least one group of checks to run")
 			firstCheck := tt.want[0]
+			// Check for this before create, otherwise it's blown away.
+			deletionTimestamp := tt.args.params.OtelCol.GetDeletionTimestamp()
 			createErr := k8sClient.Create(testContext, &tt.args.params.OtelCol)
 			if !firstCheck.validateErr(t, createErr) {
 				return
+			}
+			if deletionTimestamp != nil {
+				err := k8sClient.Delete(testContext, &tt.args.params.OtelCol, client.PropagationPolicy(metav1.DeletePropagationForeground))
+				assert.NoError(t, err)
 			}
 			req := k8sreconcile.Request{
 				NamespacedName: nsn,
@@ -523,6 +564,7 @@ func TestOpenTelemetryCollectorReconciler_Reconcile(t *testing.T) {
 			}
 			// run the next set of checks
 			for pid, updateParam := range tt.args.updates {
+				updateParam := updateParam
 				existing := v1alpha1.OpenTelemetryCollector{}
 				found, err := populateObjectIfExists(t, &existing, nsn)
 				assert.True(t, found)
@@ -642,6 +684,7 @@ func TestOpAMPBridgeReconciler_Reconcile(t *testing.T) {
 		},
 	}
 	for _, tt := range tests {
+		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			testContext := context.Background()
 			nsn := types.NamespacedName{Name: tt.args.params.OpAMPBridge.Name, Namespace: tt.args.params.OpAMPBridge.Namespace}
@@ -676,6 +719,7 @@ func TestOpAMPBridgeReconciler_Reconcile(t *testing.T) {
 			}
 			// run the next set of checks
 			for pid, updateParam := range tt.args.updates {
+				updateParam := updateParam
 				existing := v1alpha1.OpAMPBridge{}
 				found, err := populateObjectIfExists(t, &existing, nsn)
 				assert.True(t, found)
@@ -708,6 +752,43 @@ func TestOpAMPBridgeReconciler_Reconcile(t *testing.T) {
 			}
 		})
 	}
+}
+
+func TestSkipWhenInstanceDoesNotExist(t *testing.T) {
+	// prepare
+	cfg := config.New()
+	nsn := types.NamespacedName{Name: "non-existing-my-instance", Namespace: "default"}
+	reconciler := controllers.NewReconciler(controllers.Params{
+		Client: k8sClient,
+		Log:    logger,
+		Scheme: scheme.Scheme,
+		Config: cfg,
+	})
+
+	// test
+	req := k8sreconcile.Request{
+		NamespacedName: nsn,
+	}
+	_, err := reconciler.Reconcile(context.Background(), req)
+
+	// verify
+	assert.NoError(t, err)
+}
+
+func TestRegisterWithManager(t *testing.T) {
+	t.Skip("this test requires a real cluster, otherwise the GetConfigOrDie will die")
+
+	// prepare
+	mgr, err := manager.New(k8sconfig.GetConfigOrDie(), manager.Options{})
+	require.NoError(t, err)
+
+	reconciler := controllers.NewReconciler(controllers.Params{})
+
+	// test
+	err = reconciler.SetupWithManager(mgr)
+
+	// verify
+	assert.NoError(t, err)
 }
 
 func namespacedObjectName(name string, namespace string) types.NamespacedName {