+++ internal/config/main_test.go
@@ -15,15 +15,14 @@
 package config_test
 
 import (
-	"sync"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	"github.com/open-telemetry/opentelemetry-operator/internal/autodetect"
+	"github.com/open-telemetry/opentelemetry-operator/internal/autodetect/openshift"
 	"github.com/open-telemetry/opentelemetry-operator/internal/config"
-	"github.com/open-telemetry/opentelemetry-operator/pkg/autodetect"
 )
 
 func TestNewConfig(t *testing.T) {
@@ -31,78 +30,46 @@ func TestNewConfig(t *testing.T) {
 	cfg := config.New(
 		config.WithCollectorImage("some-image"),
 		config.WithCollectorConfigMapEntry("some-config.yaml"),
-		config.WithPlatform(autodetect.OpenShiftRoutesNotAvailable),
+		config.WithOpenShiftRoutesAvailability(openshift.RoutesAvailable),
 	)
 
 	// test
 	assert.Equal(t, "some-image", cfg.CollectorImage())
 	assert.Equal(t, "some-config.yaml", cfg.CollectorConfigMapEntry())
-	assert.Equal(t, autodetect.OpenShiftRoutesNotAvailable, cfg.OpenShiftRoutes())
+	assert.Equal(t, openshift.RoutesAvailable, cfg.OpenShiftRoutesAvailability())
 }
 
-func TestOnPlatformChangeCallback(t *testing.T) {
+func TestConfigChangesOnAutoDetect(t *testing.T) {
 	// prepare
-	calledBack := false
 	mock := &mockAutoDetect{
-		OpenShiftRoutesAvailabilityFunc: func() (autodetect.OpenShiftRoutesAvailability, error) {
-			return autodetect.OpenShiftRoutesAvailable, nil
+		OpenShiftRoutesAvailabilityFunc: func() (openshift.RoutesAvailability, error) {
+			return openshift.RoutesAvailable, nil
 		},
 	}
 	cfg := config.New(
 		config.WithAutoDetect(mock),
-		config.WithOnOpenShiftRoutesChangeCallback(func() error {
-			calledBack = true
-			return nil
-		}),
 	)
 
 	// sanity check
-	require.Equal(t, autodetect.OpenShiftRoutesNotAvailable, cfg.OpenShiftRoutes())
+	require.Equal(t, openshift.RoutesNotAvailable, cfg.OpenShiftRoutesAvailability())
 
 	// test
 	err := cfg.AutoDetect()
 	require.NoError(t, err)
 
 	// verify
-	assert.Equal(t, autodetect.OpenShiftRoutesAvailable, cfg.OpenShiftRoutes())
-	assert.True(t, calledBack)
-}
-
-func TestAutoDetectInBackground(t *testing.T) {
-	// prepare
-	wg := &sync.WaitGroup{}
-	wg.Add(1)
-	mock := &mockAutoDetect{
-		OpenShiftRoutesAvailabilityFunc: func() (autodetect.OpenShiftRoutesAvailability, error) {
-			wg.Done()
-			return autodetect.OpenShiftRoutesNotAvailable, nil
-		},
-	}
-	cfg := config.New(
-		config.WithAutoDetect(mock),
-		config.WithAutoDetectFrequency(500*time.Second),
-	)
-
-	// sanity check
-	require.Equal(t, autodetect.OpenShiftRoutesNotAvailable, cfg.OpenShiftRoutes())
-
-	// test
-	err := cfg.StartAutoDetect()
-	require.NoError(t, err)
-
-	// verify
-	wg.Wait()
+	assert.Equal(t, openshift.RoutesAvailable, cfg.OpenShiftRoutesAvailability())
 }
 
 var _ autodetect.AutoDetect = (*mockAutoDetect)(nil)
 
 type mockAutoDetect struct {
-	OpenShiftRoutesAvailabilityFunc func() (autodetect.OpenShiftRoutesAvailability, error)
+	OpenShiftRoutesAvailabilityFunc func() (openshift.RoutesAvailability, error)
 }
 
-func (m *mockAutoDetect) OpenShiftRoutesAvailability() (autodetect.OpenShiftRoutesAvailability, error) {
+func (m *mockAutoDetect) OpenShiftRoutesAvailability() (openshift.RoutesAvailability, error) {
 	if m.OpenShiftRoutesAvailabilityFunc != nil {
 		return m.OpenShiftRoutesAvailabilityFunc()
 	}
-	return autodetect.OpenShiftRoutesNotAvailable, nil
+	return openshift.RoutesNotAvailable, nil
 }