+++ internal/manifests/targetallocator/deployment.go
@@ -24,7 +24,7 @@ import (
 )
 
 // Deployment builds the deployment for the given instance.
-func Deployment(params manifests.Params) *appsv1.Deployment {
+func Deployment(params manifests.Params) (*appsv1.Deployment, error) {
 	name := naming.TargetAllocator(params.OtelCol.Name)
 	labels := Labels(params.OtelCol, name)
 
@@ -58,8 +58,9 @@ func Deployment(params manifests.Params) *appsv1.Deployment {
 					NodeSelector:              params.OtelCol.Spec.TargetAllocator.NodeSelector,
 					Tolerations:               params.OtelCol.Spec.TargetAllocator.Tolerations,
 					TopologySpreadConstraints: params.OtelCol.Spec.TargetAllocator.TopologySpreadConstraints,
+					Affinity:                  params.OtelCol.Spec.TargetAllocator.Affinity,
 				},
 			},
 		},
-	}
+	}, nil
 }