+++ internal/webhook/podmutation/webhookhandler_test.go
@@ -123,6 +123,7 @@ func TestShouldInjectSidecar(t *testing.T) {
 			},
 		},
 	} {
+		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			err := k8sClient.Create(context.Background(), &tt.ns)
 			require.NoError(t, err)
@@ -341,6 +342,7 @@ func TestPodShouldNotBeChanged(t *testing.T) {
 			}},
 		},
 	} {
+		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			err := k8sClient.Create(context.Background(), &tt.ns)
 			require.NoError(t, err)