+++ pkg/instrumentation/nginx_test.go
@@ -251,7 +251,7 @@ func TestInjectNginxSDK(t *testing.T) {
 					},
 				},
 			}},
-		// === Test Removal of probes  =============================
+		// === Test Removal of probes and lifecycle =============================
 		{
 			name: "Probes removed on clone init container",
 			Nginx: v1alpha1.Nginx{
@@ -267,6 +267,7 @@ func TestInjectNginxSDK(t *testing.T) {
 							ReadinessProbe: &corev1.Probe{},
 							StartupProbe:   &corev1.Probe{},
 							LivenessProbe:  &corev1.Probe{},
+							Lifecycle:      &corev1.Lifecycle{},
 						},
 					},
 				},
@@ -351,6 +352,7 @@ func TestInjectNginxSDK(t *testing.T) {
 							ReadinessProbe: &corev1.Probe{},
 							StartupProbe:   &corev1.Probe{},
 							LivenessProbe:  &corev1.Probe{},
+							Lifecycle:      &corev1.Lifecycle{},
 							Env: []corev1.EnvVar{
 								{
 									Name:  "LD_LIBRARY_PATH",